+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Moved `isValidSwissIbanNumber` and `isValidSwissSocialInsuranceNumber` to swissStandards
+
+### Fixed
+
+- `isValidSwissSocialInsuranceNumber` is now named properly
+
 ## [2.1.0] - 2025-09-03
 
 ### Added